requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-background-tasks==1.2.5
django-compat==1.0.15
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
